+++ src/config/databases/fish/species/peaceful_fish.js
@@ -9,6 +9,7 @@ const PEACEFUL_FISH = [
     visual: {
       imagePattern: "assets/fish/crucian_stalker/crucian_stalker--{level}.webp",
       uniqueLevel: 6,
+      uniqueAnomaly: "inside",
     },
 
     weatherMultipliers: {